Update inf-clojure package.
I am not sure how different the code is, but was seeing some issues at
times with a Clojure socket REPL where after a minute or so of no
output and no prompt appearing, it would output the string
'lumo.repl'.  I was not using lumo, and none of my projects were that
I could tell, but it is mentioned in inf-clojure code, so was
wondering if it was injecting it after some timeout, perhaps.