Correct small mistake in large-pkgs.py
I had not realized before that there can be multiple packages
installed simultaneously with the same package name.  Perhaps they
must have different architectures when this is the case, but do not
rely even on that -- simply allow as many duplicates with the same
name as we find in the dpkg status file.